A python script to calculate md5 checksum
MD5 File Hash in Python

The code is made to work with Python 2.7 and higher (including Python
3.x).

import hashlib

hasher = hashlib.md5()
with open('myfile.jpg', 'rb') as afile:
    buf = afile.read()
    hasher.update(buf)
print(hasher.hexdigest())
The code above calculates the MD5 digest of the file. The file is
opened in rb mode, which means that you are going to read the file in
binary mode. This is because the MD5 function needs to read the file as
a sequence of bytes. This will make sure that you can hash any type of
file, not only text files.

It is important to notice the read function. When it is called with no
arguments, like in this case, it will read all the contents of the file
and load them into memory. This is dangerous if you are not sure of the
file's size. A better version will be:

MD5 Hash for Large Files in Python

import hashlib
BLOCKSIZE = 65536
hasher = hashlib.md5()
with open('anotherfile.txt', 'rb') as afile:
    buf = afile.read(BLOCKSIZE)
    while len(buf) > 0:
        hasher.update(buf)
        buf = afile.read(BLOCKSIZE)
print(hasher.hexdigest())
If you need to use another algorithm just change the md5 call to
another supported function, e.g. SHA1:

SHA1 File Hash in Python
import hashlib
BLOCKSIZE = 65536
hasher = hashlib.sha1()
with open('anotherfile.txt', 'rb') as afile:
    buf = afile.read(BLOCKSIZE)
    while len(buf) > 0:
        hasher.update(buf)
        buf = afile.read(BLOCKSIZE)
print(hasher.hexdigest())
If you need a list of supported hash algorithms in your system use
hashlib.algorithms_available. (Only works in Python 3.2 and superior).
Finally, for another look into hashing, be sure to checkout the hashing
Python strings article.